Rag_implementation/requirements.txt
pandas
sentence-transformers
faiss-cpu
transformers
flask